pyproject.toml
[tool.poetry]
name = "agentic-kg-chatbot"
version = "0.1.0"
description = "A knowledge-graph chatbot with an agentic architecture."
authors = ["Your Name <you@example.com>"]
readme = "README.md"
packages = [{include = "crew_tools"}]

[tool.poetry.dependencies]
python = ">=3.11,<3.12"

# --- UI and Agent Frameworks (Modern Versions) ---
streamlit = "^1.33.0"
crewai = "^0.28.8"
crewai-tools = "^0.1.6"
pydantic = "^2.0.0" # We now explicitly use the modern Pydantic v2

# --- LangChain Stack (Modern and Compatible) ---
langchain = "^0.1.0"
langchain-community = "^0.0.29"
langchain-neo4j = "^0.0.4"
langchain-experimental = "^0.0.54"

# --- LLM and AI Libraries ---
ctransformers = "0.2.27"
torch = "2.2.2"
transformers = "4.39.3"
accelerate = "0.28.0"
sentence-transformers = "2.6.1"

# --- Backend and Data ---
neo4j = "5.18.0"
python-dotenv = "1.0.1"
pyvis = "0.3.2"
requests = "^2.31.0"
unstructured = {extras = ["pdf"], version = "0.12.6"}

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"